Add Components and Styles
This commit creates the basic component structure of the page: Navbar, SideBar, Page, and Footer. The layout is designed so that while not all components will be displayed necessarily, they will be present from the beginning of app production. The Navbar/SideBar/Footer will be used for site navigation, and the Page will hold the content. At present the idea is to display the Navbar/Footer only on mobile/tablet devices, while the SideBar is only meant for desktop size. As the Page component holds site content, it will always be on display. Stylesheets for each component were added to ./src/styles.

 - Create basic components: Navbar, SideBar, Page, Footer

 - Create stylesheets for each component in ./src/styles

 - - Apply background color to each component for showing sizing