Styling Responsive Components
This commit introduces rudimentary styling to set the page layout. The App component wraps children in a flexbox for vertical control of the page, and the Navbar/Footer are set vertically by flex-direction column. As the SideBar/Page are intended to flow as a row, an unnamed <div> was wrapped around them with inline styling for the moment. In addition to the basic layout, @media queries were added to NavbarStyles/FooterStyles to only show components when seen on a tablet or smaller (<768px), and the SideBarStyles set to hide when shown at that resolution.

 - Added styling (flexbox primarily) to all components

 - Added @media queries to hide SideBar below screen-width 768px

 - Added @media queries to hide Navbar/Footer above screen-width 768px