SideBar refactored to Navbar and Navbar Logic
While building the app, the layout looked compressed and cluttered. In order to generate space for the Page component, the Sidebar was set to display: none for the moment and the Navbar will handle desktop/tablet scale viewing. As the content-type Pages (Home/About/Contact) will likely not feature sufficient content to accomodate the new Page display, they were removed as links and will be placed together in the /home route. In the same vein, as education information is insufficient to merit an entire Page, the combinedLists.js experience is refactored to contain both. CardSelector is also refactored to render both CardExperience and Cardifferent components (Navbar, SideBar, Page, Footer). Keys are now generated for mapped-array components with uuid/v4

 - Move all visible components from App to Routes

 - Add state/logic for controlling active link appearance in Routes

 - Combined CardEducation/CardExperience into same CardSelector

 - Moved SideBar content into Navbar component

 - Removed About/Contact links from Navbar/SideBar (controlled in App)