ContentSelector Component Changes
As the home page is intended to be a general 'show off' area with routing to contact links and the remainder of the site, the Footer is moved to the Page component and only renders when the path is 'home' or root. Rather than write a wall of text, the background for the home page will be spheres (meant to represent electrons) orbiting elliptically around a few central self-descriptors. While accomplishing the orbit with simple rotate3D could have been performed, the result would only have the appearance of elliptical movement. To rememdy this, the addKeyFrame utility was created which takes coordinates/sizing from the ContentSelector component and creates the needed animation frames.

 - Create addKeyFrame utility for creating custom animations

 - Move Footer to Page and only to render when page is 'home'

 - Set values for ContentSelector elements that move in an orbit