Adding Options Menu to SideBar
I decided to add UI features to help with app viewing: light/dark theme, colorblind mode, and chem/dev mode. As of this commit, the Navbar contains a hamburger menu that toggles the SideBar to show/hide. While a simple 'X' was added as the obvious means to closing the SideBar, this will likely be viewed on mobile devices as well. A sibling container that occupies the remainder of the viewing window is included, and when selected (as clicking anywhere but the SideBar) is also able to toggle the view. The SideBar now includes basic functionality for an Options menu, where selected the Options toggle switch replaces the SideBar content with a series of toggle switches that control site layout: light/dark mode, colorblind mode (pure greyscale), and chem/dev mode. The useToggle hook was created and implemented in Routes.js to manage state changes.

 - Create useToggle hook for managing state in Routes.js

 - Add toggleable Options menu in the SideBar component

 - Add hamburger menu to Navbar for toggling SideBar appearance

 - Update styling for Navbar, SideBar